# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Domeye Core 查询与本地受控操作 API",
    "version": "0.1.0",
    "description": "前端查询与显式本地隔离操作的接口契约。该文件是前端原始响应类型和开发 Mock 的机器基线；W5 操作不表示生产部署。"
  },
  "servers": [{"url": "/api/v1"}],
  "paths": {
...
    "/api/v2/country-outage/capabilities/external-evidence": {
      "servers": [{"url": "/"}],
      "get": {
        "operationId": "getCountryOutageExternalEvidenceCapability",
        "description": "返回 Domeye 应用编排层的外部证据附加能力就绪态。该能力不可用不影响 RRC25 核心报告、追问和基础下载。",
        "responses": {
          "200": {
            "description": "外部证据附加能力当前就绪态",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/CountryOutageExternalEvidenceCapability"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/CountryOutageAgentUnauthenticated"},
          "503": {"$ref": "#/components/responses/CountryOutageAgentUnavailable"}
        }
      }
    },
    "/api/v2/country-outage/reports": {
      "servers": [{"url": "/"}],
      "post": {
        "operationId": "createCountryOutageAgentReport",
        "description": "由当前用户主动触发已有合法 country_outage 事件的 RRC25 固定快照报告。Python 控制面只代理到本机受限 Sidecar。",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "若提供，必须与请求体 idempotency_key 完全一致。",
            "schema": {"$ref": "#/components/schemas/CountryOutageAgentIdempotencyKey"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CountryOutageAgentCreateReportRequest"}
            }
          }
        },
        "responses": {
          "200": {"description": "幂等请求命中同一报告运行", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CountryOutageAgentCreateReportResponse"}}}},
          "202": {"description": "报告运行已排队或开始", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CountryOutageAgentCreateReportResponse"}}}},
          "400": {"$ref": "#/components/responses/CountryOutageAgentBadRequest"},
          "401": {"$ref": "#/components/responses/CountryOutageAgentUnauthenticated"},
          "403": {"$ref": "#/components/responses/CountryOutageAgentForbidden"},
          "409": {"$ref": "#/components/responses/CountryOutageAgentConflict"},
          "410": {"$ref": "#/components/responses/CountryOutageAgentExpired"},
          "429": {"$ref": "#/components/responses/CountryOutageAgentRateLimited"},
          "503": {"$ref": "#/components/responses/CountryOutageAgentUnavailable"}
        }
      }
    },
    "/api/v2/country-outage/reports/{report_id}/events": {
      "servers": [{"url": "/"}],
      "get": {
        "operationId": "streamCountryOutageAgentReportEvents",
        "description": "按递增 event id 返回受控报告、问答和会话状态；不发送模型草稿、工具返回或思考过程。支持 Last-Event-ID 重放。",
        "parameters": [
          {"name": "report_id", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^cor_[A-Za-z0-9_-]{1,124}$"}},
          {"name": "Last-Event-ID", "in": "header", "schema": {"type": "string", "pattern": "^\\d+$"}}
        ],
        "responses": {
          "200": {
            "description": "短期会话 SSE。每个 data 字段是三类受控事件之一；正式报告或答案只出现在终态事件。",
            "headers": {
              "Cache-Control": {"schema": {"type": "string"}},
              "X-Accel-Buffering": {"schema": {"type": "string", "const": "no"}}
            },
            "content": {
              "text/event-stream": {
                "schema": {"$ref": "#/components/schemas/CountryOutageAgentEvent"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/CountryOutageAgentBadRequest"},
          "401": {"$ref": "#/components/responses/CountryOutageAgentUnauthenticated"},
          "403": {"$ref": "#/components/responses/CountryOutageAgentForbidden"},
          "404": {"$ref": "#/components/responses/CountryOutageAgentNotFound"},
          "410": {"$ref": "#/components/responses/CountryOutageAgentExpired"},
          "503": {"$ref": "#/components/responses/CountryOutageAgentUnavailable"}
        }
      }
    },
    "/api/v2/country-outage/reports/{report_id}/questions": {
      "servers": [{"url": "/"}],
      "post": {
        "operationId": "askCountryOutageAgentQuestion",
        "description": "围绕原报告固定快照提交追问。默认仅使用 Domeye；只有请求显式携带本次授权时，才允许生成与 Domeye 回答严格分离的公开网络证据附录。",
        "parameters": [
          {"name": "report_id", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^cor_[A-Za-z0-9_-]{1,124}$"}},
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "若提供，必须与请求体 idempotency_key 完全一致。",
            "schema": {"$ref": "#/components/schemas/CountryOutageAgentIdempotencyKey"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CountryOutageAgentCreateQuestionRequest"}
            }
          }
        },
        "responses": {
          "200": {"description": "幂等请求命中同一问题运行", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CountryOutageAgentCreateQuestionResponse"}}}},
          "202": {"description": "问题运行已开始", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CountryOutageAgentCreateQuestionResponse"}}}},
          "400": {"$ref": "#/components/responses/CountryOutageAgentBadRequest"},
          "401": {"$ref": "#/components/responses/CountryOutageAgentUnauthenticated"},
          "403": {"$ref": "#/components/responses/CountryOutageAgentForbidden"},
          "404": {"$ref": "#/components/responses/CountryOutageAgentNotFound"},
          "409": {"$ref": "#/components/responses/CountryOutageAgentConflict"},
          "410": {"$ref": "#/components/responses/CountryOutageAgentExpired"},
          "429": {"$ref": "#/components/responses/CountryOutageAgentRateLimited"},
          "503": {"$ref": "#/components/responses/CountryOutageAgentUnavailable"}
        }
      }
    },
    "/api/v2/country-outage/runs/{run_id}/abort": {
      "servers": [{"url": "/"}],
      "post": {
        "operationId": "abortCountryOutageAgentRun",
        "description": "取消当前用户拥有的报告或问题运行；取消后不再发布迟到结果。",
        "parameters": [
          {"name": "run_id", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^run_[A-Za-z0-9_-]{1,124}$"}}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CountryOutageAgentAbortRunRequest"}}}
        },
        "responses": {
          "200": {"description": "运行已取消或已经处于终态", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CountryOutageAgentAbortRunResponse"}}}},
          "400": {"$ref": "#/components/responses/CountryOutageAgentBadRequest"},
          "401": {"$ref": "#/components/responses/CountryOutageAgentUnauthenticated"},
          "403": {"$ref": "#/components/responses/CountryOutageAgentForbidden"},
          "404": {"$ref": "#/components/responses/CountryOutageAgentNotFound"},
          "410": {"$ref": "#/components/responses/CountryOutageAgentExpired"},
          "503": {"$ref": "#/components/responses/CountryOutageAgentUnavailable"}
        }
      }
    },
    "/api/v2/country-outage/reports/{report_id}/artifacts/{artifact_format}": {
      "servers": [{"url": "/"}],
      "get": {
        "operationId": "downloadCountryOutageAgentArtifact",
        "description": "下载同一已校验报告制品生成的 Markdown 或 PDF；普通问答不进入文件。",
        "parameters": [
          {"name": "report_id", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^cor_[A-Za-z0-9_-]{1,124}$"}},
          {"name": "artifact_format", "in": "path", "required": true, "schema": {"type": "string", "enum": ["markdown", "pdf"]}}
        ],
        "responses": {
          "200": {
            "description": "报告下载制品",
            "headers": {
              "Content-Disposition": {
                "required": true,
                "schema": {"type": "string", "pattern": "^attachment;"}
              },
              "X-Artifact-Id": {
                "required": true,
                "schema": {"type": "string", "minLength": 1}
              },
              "X-Content-SHA256": {
                "required": true,
                "schema": {"type": "string", "pattern": "^[0-9a-f]{64}$"}
              }
            },
            "content": {
              "text/markdown": {"schema": {"type": "string"}},
              "application/pdf": {"schema": {"type": "string", "format": "binary"}}
            }
          },
          "400": {"$ref": "#/components/responses/CountryOutageAgentBadRequest"},
          "401": {"$ref": "#/components/responses/CountryOutageAgentUnauthenticated"},
          "403": {"$ref": "#/components/responses/CountryOutageAgentForbidden"},
          "404": {"$ref": "#/components/responses/CountryOutageAgentNotFound"},
          "409": {
            "description": "报告尚未完成，或指定格式生成失败；另一格式可能仍可下载。",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CountryOutageAgentErrorResponse"}}}
          },
          "410": {"$ref": "#/components/responses/CountryOutageAgentExpired"},
          "503": {"$ref": "#/components/responses/CountryOutageAgentUnavailable"}
        }
      }
    },
    "/api/v2/country-outage/reports/{report_id}/questions/{question_id}/artifacts/external-appendix": {
      "servers": [{"url": "/"}],
      "get": {
        "operationId": "downloadCountryOutageAgentExternalAppendix",
        "description": "在短期会话内下载绑定当前用户、基础报告固定快照和指定问题的独立 Markdown 外部来源核验附录。只有无错误的 completed 外部核验可以下载；附录不修改基础报告，普通问答不进入附录，也不提供永久历史。",
        "parameters": [
          {"name": "report_id", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^cor_[A-Za-z0-9_-]{1,124}$"}},
          {"name": "question_id", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^q_[A-Za-z0-9_-]{1,124}$"}}
        ],
        "responses": {
          "200": {
            "description": "已完成外部来源核验的独立 Markdown 附录",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {"type": "string", "const": "private, no-store"}
              },
              "X-Content-Type-Options": {
                "required": true,
                "schema": {"type": "string", "const": "nosniff"}
              },
              "Content-Disposition": {
                "required": true,
                "schema": {"type": "string", "pattern": "^attachment;"}
              },
              "X-Artifact-Id": {
                "required": true,
                "schema": {"type": "string", "pattern": "^external_appendix_[0-9a-f]{32}$"}
              },
              "X-Content-SHA256": {
                "required": true,
                "schema": {"type": "string", "pattern": "^[0-9a-f]{64}$"}
              }
            },
            "content": {
              "text/markdown": {"schema": {"type": "string"}}
            }
          },
          "400": {"$ref": "#/components/responses/CountryOutageAgentBadRequest"},
          "401": {"$ref": "#/components/responses/CountryOutageAgentUnauthenticated"},
          "403": {"$ref": "#/components/responses/CountryOutageAgentForbidden"},
          "404": {"$ref": "#/components/responses/CountryOutageAgentNotFound"},
          "409": {
            "description": "外部核验不是无错误完成态，或附录身份、来源边界不满足下载合同。",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CountryOutageAgentErrorResponse"}}}
          },
          "410": {"$ref": "#/components/responses/CountryOutageAgentExpired"},
          "503": {"$ref": "#/components/responses/CountryOutageAgentUnavailable"}
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
...
      "CountryOutageAgentBadRequest": {"description": "请求格式、标识符或幂等键不合法", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CountryOutageAgentErrorResponse"}}}},
      "CountryOutageAgentUnauthenticated": {"description": "需要有效的登录身份", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CountryOutageAgentErrorResponse"}}}},
      "CountryOutageAgentForbidden": {"description": "当前用户无事件、报告或运行访问权", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CountryOutageAgentErrorResponse"}}}},
...
      "CountryOutageAgentUnavailable": {"description": "Agent Sidecar 未配置或暂不可用", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CountryOutageAgentErrorResponse"}}}},
...
    },
    "schemas": {
...
      "CountryOutageAgentIdempotencyKey": {
        "type": "string",
        "minLength": 8,
        "maxLength": 128,
        "pattern": "^[A-Za-z0-9._:-]{8,128}$"
      },
...
      "CountryOutageExternalEvidenceCapability": {
        "oneOf": [
          {
            "type": "object",
            "required": ["schema_version", "capability", "state", "provider", "checked_at", "policy"],
            "properties": {
              "schema_version": {"type": "string", "const": "country_outage_external_evidence_capability_v1"},
              "capability": {"type": "string", "const": "external_evidence"},
              "state": {"type": "string", "const": "ready"},
              "provider": {"type": "string", "const": "managed-egress-v1"},
              "checked_at": {"type": "string", "format": "date-time"},
              "policy": {"$ref": "#/components/schemas/CountryOutageExternalEvidencePolicy"}
            },
            "additionalProperties": false
          },
          {
            "type": "object",
            "required": ["schema_version", "capability", "state", "provider", "checked_at", "policy", "reason_code"],
            "properties": {
              "schema_version": {"type": "string", "const": "country_outage_external_evidence_capability_v1"},
              "capability": {"type": "string", "const": "external_evidence"},
              "state": {"type": "string", "const": "not_configured"},
              "provider": {"type": "string", "const": "disabled"},
              "checked_at": {"type": "string", "format": "date-time"},
              "policy": {"type": "null"},
              "reason_code": {"type": "string", "minLength": 1}
            },
            "additionalProperties": false
          },
          {
            "type": "object",
            "required": ["schema_version", "capability", "state", "provider", "checked_at", "policy", "reason_code"],
            "properties": {
              "schema_version": {"type": "string", "const": "country_outage_external_evidence_capability_v1"},
              "capability": {"type": "string", "const": "external_evidence"},
              "state": {"type": "string", "const": "self_check_failed"},
              "provider": {"type": "string", "const": "managed-egress-v1"},
              "checked_at": {"type": "string", "format": "date-time"},
              "policy": {"type": "null"},
              "reason_code": {"type": "string", "minLength": 1}
            },
            "additionalProperties": false
          }
        ]
      },
...
      "CountryOutageAgentCreateReportRequest": {
        "type": "object",
        "required": ["event_reference", "publication_id", "revision", "idempotency_key"],
        "properties": {
          "event_reference": {
            "type": "string",
            "pattern": "^country_outage/\\d{4}-\\d{2}-\\d{2}[ +]\\d{2}:\\d{2}:\\d{2}/[A-Z]{2}/[1-9]\\d*/r$"
          },
          "publication_id": {"type": "string", "minLength": 1, "maxLength": 200},
          "revision": {"type": "integer", "minimum": 1},
          "idempotency_key": {"$ref": "#/components/schemas/CountryOutageAgentIdempotencyKey"}
        },
        "additionalProperties": false
      },
...
    }
  }
}